fix db config + schema
